Extracting the replayer specific utilities into a separate file for readability. (#1244)
* Fixed the spelling of replay_test file.

* Added the Activity Registration required failure scenario to replayer test suite

* Documentation change

* Removed extra file

* Extracting the replayer specific utilities into a separate file for readability.

* Pulling more helper functions

* Separated all the matching cases
